1、只安装子目录下的 rpm 安装包 2、修改 whichex 和 llex 函数 3、调整 vim 和 tmux 全局配置文件的注入和显示规则 4、调整 网卡选择菜单 中对于网卡无法取到地址情况的处理 5、对话框及安装内容的处理 6、调整 便携工具安装脚本，不重要的内容放到子目录备用，不创建软连接，调整创建软连接项目，只为文件创建软连接

+++ quick-setup/tools/install_portable_tools.sh
@@ -10,14 +10,17 @@ echo
 echo -e "\033[1mCopy the portable tools:\033[0m"
 readonly TOOLS_DIR_NAME='portable_tools'
 sudo mkdir -p /opt/$TOOLS_DIR_NAME
-sudo cp -v "$BASE_DIR/$TOOLS_DIR_NAME"/* "/opt/$TOOLS_DIR_NAME/"
+sudo cp -vR "$BASE_DIR/$TOOLS_DIR_NAME"/* "/opt/$TOOLS_DIR_NAME/"
 sudo chmod +x "/opt/$TOOLS_DIR_NAME"/*
+sudo chmod +x "/opt/$TOOLS_DIR_NAME"/**/*
 
 # 在 /usr/local/bin/ 创建 Symbolic Link
 echo
 echo -e "\033[1mCreate symlinks for portable tools:\033[0m"
 for file in /opt/"$TOOLS_DIR_NAME"/*; do
-    sudo ln -sf "$file" /usr/local/bin/"$(basename "$file")"
-    echo -e "/usr/local/bin/$(basename "$file") -> $file"
+    if [ -f "$file" ]; then
+        sudo ln -sf "$file" /usr/local/bin/"$(basename "$file")"
+        echo -e "/usr/local/bin/$(basename "$file") -> $file"
+    fi
 done
